Popup of dashboard behaviour fixed
