+++ src/app/(auth)/layout.tsx
@@ -1,33 +1,24 @@
-'use client'
+import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
 import '../globals.css'
 import { TooltipProvider } from '@/components/ui/tooltip'
-import { AuthProvider } from '@/context/authContext'
-import appwriteService from '@/appwrite/config'
-import { useEffect, useState } from 'react'
 const inter = Inter({ subsets: ['latin'] })
+
+export const metadata: Metadata = {
+  title: 'InfraBoard',
+  description: 'Visual Cloud Infrastructure Management',
+}
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode
 }>) {
-  const [authStatus, setAuthStatus] = useState(false)
-  const [loader, setLoader] = useState(true)
-  useEffect(() => {
-    appwriteService
-      .isLoggedIn()
-      .then(setAuthStatus)
-      .finally(() => setLoader(false))
-  }, [])
-
   return (
-    <AuthProvider value={{ authStatus, setAuthStatus }}>
-      {!loader && (
-        <>
-          <main className="">{children}</main>
-        </>
-      )}
-    </AuthProvider>
+    <html lang="en">
+      <body className={inter.className}>
+        <TooltipProvider delayDuration={50}>{children}</TooltipProvider>
+      </body>
+    </html>
   )
 }
