+++ src/app/page.tsx
@@ -1,37 +1,10 @@
-'use client'
-import Canvas from '@/components/canvas/canvas'
-import Header from '@/components/header/header'
-import SideBar from '@/components/sidebar/sidebar'
-import { HTML5Backend } from 'react-dnd-html5-backend'
-import { DndProvider } from 'react-dnd'
-import CodeEditor from '@/components/code_editor/codeEditor'
-import { useState } from 'react'
-import ProtectedLayout from '@/utils/protectedRoutes'
-
-interface DroppedItem {
-  subTab: {
-    title: string
-    icon?: any
-    subList: any[]
-  }
-  position: { x: number; y: number }
-}
-
-function Home() {
-  const [droppedItems, setDroppedItems] = useState<DroppedItem[]>([])
-  const handleItemDrop = (newItem: DroppedItem) => {
-    setDroppedItems((prevItems) => [...prevItems, newItem])
-  }
+import RootPage from '@/pages/home'
+const Home = () => {
   return (
     <main>
-      <DndProvider backend={HTML5Backend}>
-        <Header />
-        <SideBar />
-        <Canvas onItemDrop={handleItemDrop} droppedItems={droppedItems} />
-        <CodeEditor droppedItems={droppedItems} />
-      </DndProvider>
+      <RootPage />
     </main>
   )
 }
 
-export default ProtectedLayout(Home)
+export default Home
